Refactored RepositoryController with proper validation and delegated search to a RepositoryService
Proper validation was added by delegating it to a RepositorySearchRequest which extends Laravel's FormRequest class that allows us to specify validation rules.
We now inject a RepositoryService instance in RepositoryController, it'll allow us to separate our concerns and only have to modify the Service when changes to the search are needed.