Added a GitLab Connector, passing all tests
To add a GitLab Connector I created a ConnectorInterface with a search() method, GithubConnector and GitlabConnectors both implement the interface in a unique way, allowing flexibility in the search method which allows us to specify how to retrieve data from their respective APIs.

Both my connectors currently return empty arrays when unable to contact API server or when there are no repositories found. We could also throw an Error when unable to contact API Server and deal with it in the RepositoryService.
I decided to simply return an empty array to keep the app running and allow us to display the results fetched from 1 provider even when the other couldn't be contacted.

I also made the RepositoryController more flexible by adding a providers variable which allows us to store an array of providers.
We then loop on this array and execute RepositoryService's search method using the provider's string and the query as arguments. It allows us to individually handle exceptions and errors (We simply log the error for now).